+++ src/sample/Main.java
@@ -37,19 +37,19 @@ public class Main extends Application {
 
     //jak zrobie css to wprowadze jeszcze jedna zmienna do tej metody. narazie niech sie tak nazywa
     public static void zmiana_strony_css(ActionEvent actionEvent, Player player, FXMLLoader loader, String fxmlfile, Admin admin) throws IOException {
-        Parent home_page_parent = loader.load();
-        Scene home_page_scene = new Scene(home_page_parent);
+        Parent homePageParent = loader.load();
+        Scene homePageScene = new Scene(homePageParent);
         if (fxmlfile.equals("mainWindow")) {
-            home_page_parent.setId("pane");
+            homePageParent.setId("pane");
         }
         else if (fxmlfile.equals("rejestracja")){
-            home_page_parent.setId("rejestracja");
+            homePageParent.setId("rejestracja");
         }
         else if (fxmlfile.equals("logowanie")) {
-            home_page_parent.setId("pane");
+            homePageParent.setId("pane");
         }
-        Stage app_stage = (Stage) ((Node) actionEvent.getSource()).getScene().getWindow();
-        app_stage.setScene(home_page_scene);
-        app_stage.show();
+        Stage appStage = (Stage) ((Node) actionEvent.getSource()).getScene().getWindow();
+        appStage.setScene(homePageScene);
+        appStage.show();
     }
 }
